requirements.txt
torch
torchvision
transformers
numpy
opencv-python
matplotlib
Pillow
pandas
openpyxl
tqdm
